HttpPost Delete

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -166,13 +166,28 @@ namespace ContosoUniversity.Controllers
                 return Problem("Entity set 'SchoolContext.Students'  is null.");
             }
             var student = await _context.Students.FindAsync(id);
-            if (student != null)
+            if (student == null)
             {
-                _context.Students.Remove(student);
+                return RedirectToAction(nameof(Index));
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+
+            try
+            {
+
+                if (student != null)
+                {
+                    _context.Students.Remove(student);
+                }
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StudentExists(int id)
